move SetAddressPrefixes to init.go
