+++ tests/e2e/e2e_setup_test.go
@@ -7,12 +7,7 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/address"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
 	configurer "github.com/classic-terra/core/v2/tests/e2e/configurer"
-	"github.com/classic-terra/core/v2/types/util"
 )
 
 const (
@@ -29,37 +24,6 @@ const (
 	// Environment variable name to determine what version we are upgrading to
 	upgradeVersionEnv = "TERRA_E2E_UPGRADE_VERSION"
 )
-
-func init() {
-	SetAddressPrefixes()
-}
-
-// SetAddressPrefixes builds the Config with Bech32 addressPrefix and publKeyPrefix for accounts, validators, and consensus nodes and verifies that addreeses have correct format.
-func SetAddressPrefixes() {
-	config := sdk.GetConfig()
-	config.SetBech32PrefixForAccount(util.Bech32PrefixAccAddr, util.Bech32PrefixAccPub)
-	config.SetBech32PrefixForValidator(util.Bech32PrefixValAddr, util.Bech32PrefixValPub)
-	config.SetBech32PrefixForConsensusNode(util.Bech32PrefixConsAddr, util.Bech32PrefixConsPub)
-
-	// This is copied from the cosmos sdk v0.43.0-beta1
-	// source: https://github.com/cosmos/cosmos-sdk/blob/v0.43.0-beta1/types/address.go#L141
-	config.SetAddressVerifier(func(bytes []byte) error {
-		if len(bytes) == 0 {
-			return sdkerrors.Wrap(sdkerrors.ErrUnknownAddress, "addresses cannot be empty")
-		}
-
-		if len(bytes) > address.MaxAddrLen {
-			return sdkerrors.Wrapf(sdkerrors.ErrUnknownAddress, "address max length is %d, got %d", address.MaxAddrLen, len(bytes))
-		}
-
-		// TODO: Do we want to allow addresses of lengths other than 20 and 32 bytes?
-		if len(bytes) != 20 && len(bytes) != 32 {
-			return sdkerrors.Wrapf(sdkerrors.ErrUnknownAddress, "address length must be 20 or 32 bytes, got %d", len(bytes))
-		}
-
-		return nil
-	})
-}
 
 type IntegrationTestSuite struct {
 	suite.Suite
